requirements.txt
# PDF処理
pdf2image>=1.16.3
Pillow>=10.0.0

# OCR
pytesseract>=0.3.10
google-cloud-vision>=3.4.0

# 画像処理・特徴抽出
opencv-python>=4.8.0
numpy>=1.24.0

# テキストベクトル化
sentence-transformers>=2.2.0

# クラスタリング
scikit-learn>=1.3.0

# データベース・出力
pandas>=2.0.0

# ユーティリティ
tqdm>=4.65.0
python-dotenv>=1.0.0
